add api for identify the city name and api for identofy uset location by ip-address

+++ domain/src/main/java/com/onopry/domain/model/forecast/CurrentWeather.kt
@@ -1,9 +1,9 @@
 package com.onopry.domain.model.forecast
 
 data class CurrentWeather(
-    val temperature: Double,
+    val temperature: Int,
     val time: String,
     val weatherCode: Int,
-    val windDirection: Double,
-    val windSpeed: Double
+    val windDirection: Int,
+    val windSpeed: Int
 )
