+++ domain/src/main/java/com/onopry/domain/model/forecast/Daily.kt
@@ -1,15 +1,15 @@
 package com.onopry.domain.model.forecast
 
 data class Daily(
-    val apparentTemperatureMax: List<Double>,
-    val apparentTemperatureMin: List<Double>,
-    val precipitationHours: List<Double>,
+    val apparentTemperatureMax: List<Int>,
+    val apparentTemperatureMin: List<Int>,
+    val precipitationHours: List<Int>,
     val precipitationSum: List<Double>,
     val sunrise: List<String>,
     val sunset: List<String>,
-    val temperatureMax: List<Double>,
-    val temperatureMin: List<Double>,
+    val temperatureMax: List<Int>,
+    val temperatureMin: List<Int>,
     val time: List<String>,
     val weatherCode: List<Int>,
-    val windSpeedMax: List<Double>
+    val windSpeedMax: List<Int>
 )
